+++ packages/metadata/src/Metadata/util/toCallsOnly.ts
@@ -2,16 +2,39 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
-import { MetadataLatest } from '@polkadot/types/interfaces/metadata';
+import { FunctionMetadataLatest, MetadataLatest } from '@polkadot/types/interfaces/metadata';
 import { AnyJson, Registry } from '@polkadot/types/types';
+
+import { Option, Text, Vec } from '@polkadot/types';
+
+interface ModuleMetadataTrimmed {
+  name: Text;
+  calls: Option<Vec<FunctionMetadataLatest>>;
+}
+
+function mapCalls (registry: Registry, _calls: Option<Vec<FunctionMetadataLatest>>): Option<Vec<FunctionMetadataLatest>> {
+  const calls = _calls.unwrapOr(null);
+
+  return registry.createType(
+    'Option<Vec<FunctionMetadataLatest>>' as any,
+    calls
+      ? calls.map(({ args, documentation, name }) =>
+        registry.createType('FunctionMetadataLatest', {
+          args,
+          name,
+          documentation: documentation.map((doc) => doc.toString().trim())
+        })
+      )
+      : null
+  );
+}
 
 /** @internal */
 export default function toCallsOnly (registry: Registry, { extrinsic, modules }: MetadataLatest): AnyJson {
   return registry.createType('MetadataLatest', {
-    // FIXME, this needs typing, not any
-    modules: modules.map(({ calls, name }): any => ({
+    modules: modules.map(({ calls, name }): ModuleMetadataTrimmed => ({
       name,
-      calls
+      calls: mapCalls(registry, calls)
     })),
     extrinsic
   }).toJSON();