+++ packages/types/src/create/sanitize.ts
@@ -32,12 +32,10 @@ const mappings: Mapper[] = [
   alias(['Lookup::Target'], 'LookupTarget'),
   // HACK duplication between contracts & primitives, however contracts prefixed with exec
   alias(['exec::StorageKey'], 'ContractStorageKey'),
-  // alias for internal module mappings
-  alias(['exec', 'grandpa', 'marker', 'session', 'slashing'].map((s) => `${s}::`), ''),
   // flattens tuples with one value, `(AccountId)` -> `AccountId`
   flattenSingleTuple(),
   // converts ::Type to Type, <T as Trait<I>>::Proposal -> Proposal
-  removeColonPrefix()
+  removeColons()
 ];
 
 // given a starting index, find the closing >
@@ -98,9 +96,27 @@ export function flattenSingleTuple (): Mapper {
   };
 }
 
-export function removeColonPrefix (): Mapper {
+export function removeColons (): Mapper {
   return (value: string): string => {
-    return value.replace(/^::/, '');
+    let index = 0;
+
+    while (index !== -1) {
+      index = value.indexOf('::');
+
+      if (index === 0) {
+        value = value.substr(2);
+      } else if (index !== -1) {
+        let start = index;
+
+        while (start !== -1 && !BOX_PRECEDING.includes(value[start])) {
+          start--;
+        }
+
+        value = `${value.substr(0, start + 1)}${value.substr(index + 2)}`;
+      }
+    }
+
+    return value;
   };
 }
 
@@ -155,7 +171,7 @@ export function removeTraits (): Mapper {
       // remove all whitespaces
       .replace(/\s/g, '')
       // anything `T::<type>` to end up as `<type>`
-      .replace(/(T|Self|wasm)::/g, '')
+      .replace(/(T|Self)::/g, '')
       // replace `<T as Trait>::` (whitespaces were removed above)
       .replace(/<(T|Self)asTrait>::/g, '')
       // replace `<T as something::Trait>::` (whitespaces were removed above)
@@ -163,9 +179,7 @@ export function removeTraits (): Mapper {
       // replace <Lookup as StaticLookup>
       .replace(/<LookupasStaticLookup>/g, 'Lookup')
       // replace `<...>::Type`
-      .replace(/::Type/g, '')
-      // `sr_std::marker::`
-      .replace(/(sp_std|sr_std|rstd)::/g, '');
+      .replace(/::Type/g, '');
   };
 }