+++ packages/types/src/interfaces/consensus/types.ts
@@ -1,9 +1,13 @@
 // Auto-generated via `yarn polkadot-types-from-defs`, do not edit
 /* eslint-disable */
 
+import { U8aFixed } from '@polkadot/types/codec';
 import { AccountId } from '@polkadot/types/interfaces/runtime';
 
 /** @name AuthorityId */
 export interface AuthorityId extends AccountId {}
+
+/** @name RawVRFOutput */
+export interface RawVRFOutput extends U8aFixed {}
 
 export type PHANTOM_CONSENSUS = 'consensus';