+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # 1.9.0-beta.x
 
+- Tx signing via `signAndSend` & `signAsync` now tries to use the last finalized block (depending on finalized drift)
 - Add support for `api.rpc.childstate.*` Substrate RPCs, older nodes with still populate `api.rpc.state.*` with the child methods
-- Additional cleanups of democracy derives, including support for new Polkadot/Substrate vote retrievals
+- Adjust `Error.message` on RPC failures to not include the (already known) method signature
+- Additional cleanups and rework of democracy derives, including support for new Polkadot/Substrate vote retrievals
+- Internal adjustment of interface types to align with modules (no external impact via `@polkadot/types/interfaces`)
 
 # 1.8.1 Mar 22, 2020